--- src/ui/layouts/Delegation.tsx
@@ -1,216 +1,0 @@
-import { useState } from 'react';
-import { Button } from '../lib';
-import { DelegateCard } from '../components/delegation/DelegateCard';
-import { TrackSelect } from '../components/delegation/TrackSelect.js';
-import { AddIcon, ChevronDownIcon } from '../icons';
-import { useDelegation } from '../../contexts/Delegation';
-import SectionTitle from '../components/SectionTitle';
-import ProgressStepper from '../components/ProgressStepper.js';
-import { ConnectedState, Delegate, State } from '../../lifecycle/types.js';
-import {
-  useAppLifeCycle,
-  filterOngoingReferenda,
-  flattenAllTracks,
-  extractDelegatedTracks,
-} from '../../lifecycle';
-import { ReferendumOngoing } from '../../types';
-import Headline from '../components/Headline';
-import { DelegateModal } from '../components/delegation/delegateModal/Delegate';
-import { AddDelegateModal } from '../components/delegation/delegateModal/AddDelegateModal';
-import { DelegatesBar } from '../components/DelegatesBar';
-import { ActiveDelegates } from '../components/ActiveDelegates';
-
-export const DelegateSection = ({
-  state,
-  delegates,
-}: {
-  state: State;
-  delegates: Delegate[];
-}) => {
-  const [search, setSearch] = useState<string>();
-  const [addAddressVisible, setAddAddressVisible] = useState(false);
-  const [delegateVisible, setDelegateVisible] = useState(false);
-  const [delegate, setDelegate] = useState('');
-  const { selectedTrackIndexes } = useDelegation();
-  const allTracks = flattenAllTracks(state.tracks);
-  const selectedTracks = Array.from(selectedTrackIndexes.entries()).map(
-    ([id]) => allTracks.get(id)!
-  );
-
-  return (
-    <>
-      <div className="mb-48 flex w-full flex-col gap-16">
-        <SectionTitle
-          title="Browse Delegates"
-          description={
-            <span>
-              There are currently <b>{delegates.length}</b> delegates.
-            </span>
-          }
-        >
-          <ProgressStepper step={1} />
-        </SectionTitle>
-        <div className=" flex flex-col gap-4 ">
-          <div className="sticky top-44 flex w-full flex-col items-center justify-between gap-4 bg-bg-default/80 px-3 py-3 backdrop-blur-md md:flex-row lg:px-8">
-            <div className="flex w-full flex-row items-center justify-between gap-4">
-              <Button
-                variant="ghost"
-                className="w-fit"
-                onClick={() => setAddAddressVisible(true)}
-              >
-                <AddIcon />
-                <div className="whitespace-nowrap">Add address</div>
-              </Button>
-              <input
-                placeholder="Search"
-                className="w-full self-stretch rounded-lg bg-[#ebeaea] px-4 py-2 text-left text-sm text-black opacity-70 lg:w-fit"
-                onChange={(event) => setSearch(event.target.value)}
-              />
-            </div>
-          </div>
-          {state.customDelegates?.length > 0 && (
-            <div className="flex flex-col gap-4 px-3 lg:px-8">
-              <div className="text-sm">Added manually</div>
-              <div className="grid grid-cols-1 place-items-center  gap-2 md:grid-cols-2 lg:grid-cols-3 lg:gap-4">
-                {state.customDelegates.map((delegate, idx) => (
-                  <DelegateCard
-                    key={idx}
-                    delegate={delegate}
-                    state={state}
-                    variant="some"
-                  />
-                ))}
-              </div>
-              <div className="text-sm">Public Delegates</div>
-            </div>
-          )}
-          <div className="grid grid-cols-1 place-items-center gap-2 px-3  md:grid-cols-2 lg:grid-cols-3 lg:gap-4 lg:px-8">
-            {delegates
-              .filter((delegate) =>
-                search
-                  ? delegate.name?.toLowerCase().includes(search.toLowerCase())
-                  : true
-              )
-              .map((delegate, idx) => (
-                <DelegateCard
-                  key={idx}
-                  delegate={delegate}
-                  state={state}
-                  variant="some"
-                />
-              ))}
-          </div>
-        </div>
-      </div>
-      <AddDelegateModal
-        open={addAddressVisible}
-        onClose={() => setAddAddressVisible(false)}
-        onAddressValidated={(address) => {
-          setDelegate(address);
-          setAddAddressVisible(false);
-          setDelegateVisible(true);
-        }}
-      />
-      {delegateVisible && (
-        <DelegateModal
-          open={delegateVisible}
-          onClose={() => setDelegateVisible(false)}
-          delegate={delegate}
-          selectedTracks={selectedTracks}
-        />
-      )}
-    </>
-  );
-};
-
-function exportReferenda(state: State): Map<number, ReferendumOngoing> {
-  if (state.type === 'ConnectedState') {
-    return filterOngoingReferenda(state.chain.referenda);
-  }
-  return new Map();
-}
-
-function DelegationPanelContent({
-  state,
-  delegates,
-}: {
-  state: State;
-  delegates: Delegate[];
-}): JSX.Element {
-  const network = (state as ConnectedState).network;
-
-  const { selectedTrackIndexes, sectionRefs, scrollToSection } =
-    useDelegation();
-
-  const delegatesWithTracks = extractDelegatedTracks(state);
-
-  return (
-    <>
-      {delegatesWithTracks.size ? (
-        <div className="flex snap-start flex-col items-center">
-          <span className="px-3 font-unbounded text-h4">
-            Expand your delegation
-          </span>
-          <p className="text-body">
-            You can always delegate undelegated tracks without locking any more
-            tokens.
-          </p>
-          <ChevronDownIcon className="mt-4" />
-        </div>
-      ) : (
-        <div className="flex snap-start flex-col items-center">
-          <span className="px-3 font-unbounded text-h4">
-            Set up your delegation preferences
-          </span>
-          <p className="text-body">
-            First, select the tracks you would like to delegate, then pick the
-            address you&apos;d like to delegate to.
-          </p>
-        </div>
-      )}
-
-      <TrackSelect
-        network={network}
-        details={state.details}
-        referenda={exportReferenda(state)}
-        tracks={state.tracks}
-        delegateHandler={() => scrollToSection('delegation')}
-      />
-      {selectedTrackIndexes.size > 0 && (
-        <div className="w-full" ref={sectionRefs.get('delegation')}>
-          <DelegateSection state={state} delegates={delegates} />
-        </div>
-      )}
-    </>
-  );
-}
-
-export function DelegationPanel() {
-  const { state } = useAppLifeCycle();
-  const { delegates } = state;
-  const { sectionRefs } = useDelegation();
-
-  // A map of delegates with asociated tracks. Empty if no tracks are currently delegated.
-  const delegatesWithTracks = extractDelegatedTracks(state);
-
-  // If user has some active delegation,
-  return (
-    <main
-      className="flex w-full flex-auto snap-y flex-col items-center justify-start gap-8 pt-14 md:pt-20 lg:gap-16"
-      ref={sectionRefs.get('top')}
-    >
-      {delegatesWithTracks.size ? (
-        <ActiveDelegates
-          delegatesWithTracks={delegatesWithTracks}
-          state={state}
-        />
-      ) : (
-        <>
-          <Headline />
-          <DelegatesBar delegates={delegates} state={state} />
-        </>
-      )}
-      <DelegationPanelContent delegates={delegates} state={state} />
-    </main>
-  );
-}
